chore(cz-git): rename folder name
