+++ packages/cz-git/src/generator/index.ts
@@ -1,3 +1,3 @@
-export * from "./build";
+export * from "./message";
 export * from "./option";
 export * from "./question";
